requirement.txt
streamlit
google-generativeai
python-dotenv
transformers
torch